refactor: simplify String length method body
